# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NiazFinder API",
    "version": "0.2.x-unstable",
    "description": "Generated skeleton of the Next.js App Router API surface. Methods and auth requirements are statically detected; only operations with explicit request/response schemas are stable. Regenerate with `npm run docs:openapi`.",
    "license": {
      "name": "MIT",
      "url": "https://github.com/sasanappstore2/NiazFinder-main/blob/main/LICENSE"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local dev (`npm run dev`)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "User token from phone-OTP login (AuthToken)."
      },
      "internalSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "x-internal-secret",
        "description": "INTERNAL_API_SECRET for worker/cron routes. Constant-time compare, fail-closed."
      }
    }
  },
  "paths": {
...
    "/api/business/me/categories": {
      "get": {
        "tags": [
          "business"
        ],
        "summary": "GET /api/business/me/categories",
        "description": "Authenticated route (Bearer token). Elevated operations additionally require ADMIN / SUPER_ADMIN role.",
        "responses": {
          "200": {
            "description": "Success (schema not yet documented — see route source)."
          },
          "400": {
            "description": "Validation error."
          },
          "401": {
            "description": "Unauthenticated."
          },
          "403": {
            "description": "Forbidden."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "business"
        ],
        "summary": "PATCH /api/business/me/categories",
        "description": "Authenticated route (Bearer token). Elevated operations additionally require ADMIN / SUPER_ADMIN role.",
        "responses": {
          "200": {
            "description": "Success (schema not yet documented — see route source)."
          },
          "400": {
            "description": "Validation error."
          },
          "401": {
            "description": "Unauthenticated."
          },
          "403": {
            "description": "Forbidden."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/business/me/offers": {
      "get": {
        "tags": [
          "business"
        ],
        "summary": "GET /api/business/me/offers",
        "description": "Authenticated route (Bearer token). Elevated operations additionally require ADMIN / SUPER_ADMIN role.",
        "responses": {
          "200": {
            "description": "Success (schema not yet documented — see route source)."
          },
          "400": {
            "description": "Validation error."
          },
          "401": {
            "description": "Unauthenticated."
          },
          "403": {
            "description": "Forbidden."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "business"
        ],
        "summary": "POST /api/business/me/offers",
        "description": "Authenticated route (Bearer token). Elevated operations additionally require ADMIN / SUPER_ADMIN role.",
        "responses": {
          "200": {
            "description": "Success (schema not yet documented — see route source)."
          },
          "400": {
            "description": "Validation error."
          },
          "401": {
            "description": "Unauthenticated."
          },
          "403": {
            "description": "Forbidden."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/business/me/onboarding": {
      "post": {
        "tags": [
          "business"
        ],
        "summary": "POST /api/business/me/onboarding",
        "description": "No server-side auth call detected statically — verify before treating as public.",
        "responses": {
          "200": {
            "description": "Success (schema not yet documented — see route source)."
          },
          "400": {
            "description": "Validation error."
          },
          "401": {
            "description": "Unauthenticated."
          },
          "403": {
            "description": "Forbidden."
          }
        }
      }
    },
...
    "/api/chat/{conversationId}/mute": {
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "GET /api/chat/[conversationId]/mute",
        "description": "Authenticated route (Bearer token). Elevated operations additionally require ADMIN / SUPER_ADMIN role.",
        "responses": {
          "200": {
            "description": "Success (schema not yet documented — see route source)."
          },
          "400": {
            "description": "Validation error."
          },
          "401": {
            "description": "Unauthenticated."
          },
          "403": {
            "description": "Forbidden."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "chat"
        ],
        "summary": "PATCH /api/chat/[conversationId]/mute",
        "description": "Authenticated route (Bearer token). Elevated operations additionally require ADMIN / SUPER_ADMIN role.",
        "responses": {
          "200": {
            "description": "Success (schema not yet documented — see route source)."
          },
          "400": {
            "description": "Validation error."
          },
          "401": {
            "description": "Unauthenticated."
          },
          "403": {
            "description": "Forbidden."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  }
}